+++ trucks/src/main/java/com/truckplatoon/trucks/repository/TruckRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.truckplatoon.trucks.model.Truck;
 
-public interface TruckRepository extends JpaRepository<Truck, Long> {
+public interface TruckRepository extends JpaRepository<Truck, Integer> {
 
 }
